+++ src/Pim/Bundle/ImportExportBundle/Processor/CategoryProcessor.php
@@ -52,7 +52,11 @@ class CategoryProcessor extends AbstractEntityProcessor
     {
         return array(
             'circularRefsChecked' => array(
-                'type' => 'switch',
+                'type'    => 'switch',
+                'options' => array(
+                    'label' => 'pim_import_export.import.circularRefsChecked.label',
+                    'help'  => 'pim_import_export.import.circularRefsChecked.help'
+                )
             ),
         );
     }